Extract image layer snapshot infrastructure
